Put all Minitest::Spec::DSL tests in describe blocks
This is pretty grim - sorry!

It has transpired that some of the tests of this app aren't actually
running in CI. This is because the parallel test runner is not
distinguishing them as different from other tests. The tests I noticed
not to be running were in test/unit/services/asset_manager/attachment_updater/attachment_redirect_url_updater_test.rb
and I established they weren't running after pondering for a while how
on earth they were ever passing in CI - of course they weren't being
run.

The reason for this is that they were identified based on their context
blocks which are the same between multiple tests. I figure the best way
to resolve this is to set describe blocks on all Minitest::Spec::DSL
tests so that they are all distinguished.

Of course, the underlying problem of all of this is that for some reason
we have multiple approaches to testing Ruby in this app. In my opinion
the cleanest solution to this is to rewrite all of the
Minitest::Spec::DSL tests to use the standard Minitest approach.
