+++ test/unit/services/asset_manager/attachment_updater/attachment_access_limited_worker_test.rb
@@ -3,93 +3,95 @@ require "test_helper"
 class AssetManager::AttachmentAccessLimitedUpdaterTest < ActiveSupport::TestCase
   extend Minitest::Spec::DSL
 
-  let(:updater) { AssetManager::AttachmentUpdater }
-  let(:attachment_data) { attachment.attachment_data }
-  let(:update_worker) { mock("asset-manager-update-asset-worker") }
+  describe AssetManager::AttachmentUpdater::AccessLimitedUpdates do
+    let(:updater) { AssetManager::AttachmentUpdater }
+    let(:attachment_data) { attachment.attachment_data }
+    let(:update_worker) { mock("asset-manager-update-asset-worker") }
 
-  around do |test|
-    AssetManager.stub_const(:AssetUpdater, update_worker) do
-      test.call
-    end
-  end
-
-  context "when attachment's attachable is access limited and the attachment is not a PDF" do
-    let(:sample_rtf) { File.open(fixture_path.join("sample.rtf")) }
-    let(:attachment) { FactoryBot.create(:file_attachment, file: sample_rtf) }
-
-    before do
-      AttachmentData.stubs(:find_by).with(id: attachment_data.id).returns(attachment_data)
-
-      access_limited_object = stub("access-limited-object")
-      AssetManagerAccessLimitation.stubs(:for).with(access_limited_object).returns(%w[user-uid])
-
-      attachment_data.stubs(:access_limited?).returns(true)
-      attachment_data.stubs(:access_limited_object).returns(access_limited_object)
+    around do |test|
+      AssetManager.stub_const(:AssetUpdater, update_worker) do
+        test.call
+      end
     end
 
-    it "updates the access limited state of the asset" do
-      update_worker.expects(:call)
-        .with(attachment_data, attachment.file.asset_manager_path, "access_limited" => %w[user-uid])
+    context "when attachment's attachable is access limited and the attachment is not a PDF" do
+      let(:sample_rtf) { File.open(fixture_path.join("sample.rtf")) }
+      let(:attachment) { FactoryBot.create(:file_attachment, file: sample_rtf) }
 
-      updater.call(attachment_data, access_limited: true)
-    end
-  end
+      before do
+        AttachmentData.stubs(:find_by).with(id: attachment_data.id).returns(attachment_data)
 
-  context "when attachment's attachable is access limited and the attachment is a PDF" do
-    let(:simple_pdf) { File.open(fixture_path.join("simple.pdf")) }
-    let(:attachment) { FactoryBot.create(:file_attachment, file: simple_pdf) }
+        access_limited_object = stub("access-limited-object")
+        AssetManagerAccessLimitation.stubs(:for).with(access_limited_object).returns(%w[user-uid])
 
-    before do
-      AttachmentData.stubs(:find_by).with(id: attachment_data.id).returns(attachment_data)
+        attachment_data.stubs(:access_limited?).returns(true)
+        attachment_data.stubs(:access_limited_object).returns(access_limited_object)
+      end
 
-      access_limited_object = stub("access-limited-object")
-      AssetManagerAccessLimitation.stubs(:for).with(access_limited_object).returns(%w[user-uid])
+      it "updates the access limited state of the asset" do
+        update_worker.expects(:call)
+          .with(attachment_data, attachment.file.asset_manager_path, "access_limited" => %w[user-uid])
 
-      attachment_data.stubs(:access_limited?).returns(true)
-      attachment_data.stubs(:access_limited_object).returns(access_limited_object)
-    end
-
-    it "updates the access limited state of the asset and it's thumbnail" do
-      update_worker.expects(:call)
-        .with(attachment_data, attachment.file.asset_manager_path, "access_limited" => %w[user-uid])
-      update_worker.expects(:call)
-        .with(attachment_data, attachment.file.thumbnail.asset_manager_path, "access_limited" => %w[user-uid])
-
-      updater.call(attachment_data, access_limited: true)
-    end
-  end
-
-  context "when attachment's attachable is not access limited and the attachment is not a PDF" do
-    let(:sample_rtf) { File.open(fixture_path.join("sample.rtf")) }
-    let(:attachment) { FactoryBot.create(:file_attachment, file: sample_rtf) }
-
-    before do
-      attachment_data.stubs(:access_limited?).returns(false)
+        updater.call(attachment_data, access_limited: true)
+      end
     end
 
-    it "updates the asset to have an empty access_limited array" do
-      update_worker.expects(:call)
-        .with(attachment_data, attachment.file.asset_manager_path, "access_limited" => [])
+    context "when attachment's attachable is access limited and the attachment is a PDF" do
+      let(:simple_pdf) { File.open(fixture_path.join("simple.pdf")) }
+      let(:attachment) { FactoryBot.create(:file_attachment, file: simple_pdf) }
 
-      updater.call(attachment_data, access_limited: true)
+      before do
+        AttachmentData.stubs(:find_by).with(id: attachment_data.id).returns(attachment_data)
+
+        access_limited_object = stub("access-limited-object")
+        AssetManagerAccessLimitation.stubs(:for).with(access_limited_object).returns(%w[user-uid])
+
+        attachment_data.stubs(:access_limited?).returns(true)
+        attachment_data.stubs(:access_limited_object).returns(access_limited_object)
+      end
+
+      it "updates the access limited state of the asset and it's thumbnail" do
+        update_worker.expects(:call)
+          .with(attachment_data, attachment.file.asset_manager_path, "access_limited" => %w[user-uid])
+        update_worker.expects(:call)
+          .with(attachment_data, attachment.file.thumbnail.asset_manager_path, "access_limited" => %w[user-uid])
+
+        updater.call(attachment_data, access_limited: true)
+      end
     end
-  end
 
-  context "when attachment's attachable is not access limited and the attachment is a PDF" do
-    let(:simple_pdf) { File.open(fixture_path.join("simple.pdf")) }
-    let(:attachment) { FactoryBot.create(:file_attachment, file: simple_pdf) }
+    context "when attachment's attachable is not access limited and the attachment is not a PDF" do
+      let(:sample_rtf) { File.open(fixture_path.join("sample.rtf")) }
+      let(:attachment) { FactoryBot.create(:file_attachment, file: sample_rtf) }
 
-    before do
-      attachment_data.stubs(:access_limited?).returns(false)
+      before do
+        attachment_data.stubs(:access_limited?).returns(false)
+      end
+
+      it "updates the asset to have an empty access_limited array" do
+        update_worker.expects(:call)
+          .with(attachment_data, attachment.file.asset_manager_path, "access_limited" => [])
+
+        updater.call(attachment_data, access_limited: true)
+      end
     end
 
-    it "updates the asset to have an empty access_limited array" do
-      update_worker.expects(:call)
-        .with(attachment_data, attachment.file.asset_manager_path, "access_limited" => [])
-      update_worker.expects(:call)
-        .with(attachment_data, attachment.file.thumbnail.asset_manager_path, "access_limited" => [])
+    context "when attachment's attachable is not access limited and the attachment is a PDF" do
+      let(:simple_pdf) { File.open(fixture_path.join("simple.pdf")) }
+      let(:attachment) { FactoryBot.create(:file_attachment, file: simple_pdf) }
 
-      updater.call(attachment_data, access_limited: true)
+      before do
+        attachment_data.stubs(:access_limited?).returns(false)
+      end
+
+      it "updates the asset to have an empty access_limited array" do
+        update_worker.expects(:call)
+          .with(attachment_data, attachment.file.asset_manager_path, "access_limited" => [])
+        update_worker.expects(:call)
+          .with(attachment_data, attachment.file.thumbnail.asset_manager_path, "access_limited" => [])
+
+        updater.call(attachment_data, access_limited: true)
+      end
     end
   end
 end
