+++ test/unit/services/asset_manager/attachment_updater_test.rb
@@ -3,14 +3,16 @@ require "test_helper"
 class AssetManager::AttachmentUpdaterTest < ActiveSupport::TestCase
   extend Minitest::Spec::DSL
 
-  let(:subject) { AssetManager::AttachmentUpdater }
-  let(:file) { File.open(fixture_path.join("sample.rtf")) }
-  let(:attachment) { FactoryBot.create(:file_attachment, file: file) }
-  let(:attachment_data) { attachment.attachment_data }
+  describe AssetManager::AttachmentUpdater do
+    let(:subject) { AssetManager::AttachmentUpdater }
+    let(:file) { File.open(fixture_path.join("sample.rtf")) }
+    let(:attachment) { FactoryBot.create(:file_attachment, file: file) }
+    let(:attachment_data) { attachment.attachment_data }
 
-  it "groups updates together" do
-    AssetManager::AssetUpdater.expects(:call).once
+    it "groups updates together" do
+      AssetManager::AssetUpdater.expects(:call).once
 
-    subject.call(attachment_data, redirect_url: true, draft_status: true)
+      subject.call(attachment_data, redirect_url: true, draft_status: true)
+    end
   end
 end
