+++ test/unit/workers/asset_manager_attachment_redirect_url_update_worker_test.rb
@@ -3,13 +3,15 @@ require "test_helper"
 class AssetManagerAttachmentRedirectUrlUpdateWorkerTest < ActiveSupport::TestCase
   extend Minitest::Spec::DSL
 
-  let(:file) { File.open(fixture_path.join("sample.rtf")) }
-  let(:attachment) { FactoryBot.create(:file_attachment, file: file) }
-  let(:attachment_data) { attachment.attachment_data }
-  let(:worker) { AssetManagerAttachmentRedirectUrlUpdateWorker.new }
+  describe AssetManagerAttachmentRedirectUrlUpdateWorker do
+    let(:file) { File.open(fixture_path.join("sample.rtf")) }
+    let(:attachment) { FactoryBot.create(:file_attachment, file: file) }
+    let(:attachment_data) { attachment.attachment_data }
+    let(:worker) { AssetManagerAttachmentRedirectUrlUpdateWorker.new }
 
-  it "calls AssetManager::AttachmentRedirectUrlUpdater" do
-    AssetManager::AttachmentUpdater.expects(:call).with(attachment_data, redirect_url: true)
-    worker.perform(attachment_data.id)
+    it "calls AssetManager::AttachmentRedirectUrlUpdater" do
+      AssetManager::AttachmentUpdater.expects(:call).with(attachment_data, redirect_url: true)
+      worker.perform(attachment_data.id)
+    end
   end
 end
